+++ src/main/kotlin/com/magidc/ideavim/anyobject/handlers/AnyBlockCommentHandler.kt
@@ -1,9 +1,8 @@
 package com.magidc.ideavim.anyobject.handlers
 
 import com.magidc.ideavim.anyobject.handlers.base.DelimiterHandler
-import com.magidc.ideavim.anyobject.handlers.base.TextBasedHandlerFactory
 
-class AnyBlockCommentHandlers : TextBasedHandlerFactory {
+class AnyBlockCommentHandler : DelimiterHandler(false, delimiters) {
     companion object {
         private val delimiters = listOf(
             "/**" to "*/",          // Java
@@ -16,13 +15,5 @@ class AnyBlockCommentHandlers : TextBasedHandlerFactory {
             "#'" to "'",            // R (roxygen comments, though not exactly block comments)
             "{-" to "-}",           // Haskell
         )
-    }
-
-    override fun getInnerHandler(): DelimiterHandler {
-        return DelimiterHandler(isInner = true, sameLine = false, delimiterPairs = delimiters)
-    }
-
-    override fun getOuterHandler(size: Int): DelimiterHandler {
-        return DelimiterHandler(isInner = false, sameLine = false, delimiterPairs = delimiters)
     }
 }